fix(web-serial): break shared circular dependency for text meta
Move TextFileMeta and editor size limits into web-serial so shared can
keep depending on web-serial without a build graph cycle.
